Got DELETE working, minor cleanups all around
