+++ beerweb/ui/app/actions/api/index.js
@@ -1,0 +1,7 @@
+export * from './BreweryApiActions';
+export * from './BeerApiActions';
+export * from './KegApiActions';
+
+export function shouldFetchArrayType(arrayFromState) {
+  return !!(!arrayFromState || arrayFromState.length === 0);
+}
