+++ beerweb/ui/app/actions/constants/ActionTypes.js
@@ -1,9 +1,7 @@
-/*
- * action types
- */
-export const SET_BREWERY_VISIBILITY_FILTER = 'SET_BREWERY_VISIBILITY_FILTER';
-export const REQUEST_BREWERIES = 'REQUEST_BREWERIES';
-export const RECEIVE_BREWERIES = 'RECEIVE_BREWERIES';
+export const GET_BREWERIES_SUCCESS = 'GET_BREWERIES_SUCCESS';
+export const POST_BREWERIES_SUCCESS = 'POST_BREWERIES_SUCCESS';
+export const PUT_BREWERY_SUCCESS = 'PUT_BREWERY_SUCCESS';
+export const DELETE_BREWERY_SUCCESS = 'DELETE_BREWERY_SUCCESS';
 export const REQUEST_BEERS = 'REQUEST_BEERS';
 export const RECEIVE_BEERS = 'RECEIVE_BEERS';
 export const REQUEST_BEER_STYLES = 'REQUEST_BEER_STYLES';
@@ -13,6 +11,4 @@ export const RECEIVE_KEGS = 'RECEIVE_KEGS';
 export const ADD_BREWERY = 'ADD_BREWERY';
 export const UPDATE_BREWERY_NAME_LOCALLY = 'UPDATE_BREWERY_NAME_LOCALLY';
 export const PAGE_UNMOUNTED = 'PAGE_UNMOUNTED';
-export const POST_BREWERIES_SUCCESS = 'POST_BREWERIES_SUCCESS';
-export const PUT_BREWERY_SUCCESS = 'PUT_BREWERY_SUCCESS';
 export const SET_ROW_AS_EDITABLE = 'SET_ROW_AS_EDITABLE';
