+++ beerweb/ui/app/reducers/BreweriesReducer.js
@@ -1,12 +1,11 @@
-import {VisibilityFilters} from "../actions/index";
 import {
-  RECEIVE_BREWERIES,
+  GET_BREWERIES_SUCCESS,
   ADD_BREWERY,
   UPDATE_BREWERY_NAME_LOCALLY,
   PAGE_UNMOUNTED,
-  SET_BREWERY_VISIBILITY_FILTER,
   SET_ROW_AS_EDITABLE,
-  PUT_BREWERY_SUCCESS
+  PUT_BREWERY_SUCCESS,
+  DELETE_BREWERY_SUCCESS
 } from "../actions/constants/ActionTypes";
 
 
@@ -18,6 +17,56 @@ function getEmptyBrewery() {
   };
 }
 
+var removeBreweryFromState = function (state, action) {
+  let newState = Object.assign({}, state);
+  newState.items = state.items.filter(brewery=>brewery.brewery_id !== action.breweryId);
+  return newState;
+};
+
+var replaceStateBreweries = function (state, action) {
+  return Object.assign({}, state, {
+    items: action.breweries
+  });
+};
+var addSingleNewBreweryToState = function (state) {
+  return Object.assign({}, state, {
+    isUiDirty: true,
+    newItems: [...state.newItems, getEmptyBrewery()]
+  });
+};
+var updateBreweryNameInState = function (state, action) {
+  let newState = Object.assign({}, state);
+
+  if (action.isExistingBrewery) {
+    newState.items.find(item=>item.brewery_id === action.breweryId).name = action.name;
+  } else {
+    newState.newItems.find(item=>item.brewery_id === action.breweryId).name = action.name;
+  }
+
+  newState.isUserInputValid =
+    newState.items.every(item=>item.name && item.name.length > 0) &&
+    newState.newItems.every(item=>item.name && item.name.length > 0);
+
+  return newState;
+};
+var clearAnyUnsavedBreweriesFromState = function (state) {
+  return Object.assign({}, state, {
+    newItems: [],
+    isUiDirty: false
+  });
+};
+
+var markBreweryAsEditable = function (state, action) {
+  let stateCopy = Object.assign({}, state);
+  stateCopy.items.find(item=>item.brewery_id === action.rowId).isEditable = true;
+  return stateCopy;
+};
+
+var clearBreweriesFromState = function (state) {
+  return Object.assign({}, state, {
+    items: []
+  });
+};
 
 export function breweries(state = {
   items: [],
@@ -26,55 +75,22 @@ export function breweries(state = {
   isUserInputValid: false
 }, action) {
   switch (action.type) {
-    case RECEIVE_BREWERIES:
-      return Object.assign({}, state, {
-        items: action.breweries
-      });
+    case GET_BREWERIES_SUCCESS:
+      return replaceStateBreweries(state, action);
     case ADD_BREWERY:
-      return Object.assign({}, state, {
-        isUiDirty: true,
-        newItems: [...state.newItems, getEmptyBrewery()]
-      });
+      return addSingleNewBreweryToState(state);
     case UPDATE_BREWERY_NAME_LOCALLY:
-      let newState = Object.assign({}, state);
-
-      if (action.isExistingBrewery) {
-        newState.items.find(item=>item.brewery_id === action.breweryId).name = action.name;
-      }
-      else {
-        newState.newItems.find(item=>item.brewery_id === action.breweryId).name = action.name;
-      }
-
-      console.log(JSON.stringify(state));
-      newState.isUserInputValid =
-        newState.items.every(item=>item.name && item.name.length > 0) &&
-        newState.newItems.every(item=>item.name && item.name.length > 0);
-
-      return newState;
+      return updateBreweryNameInState(state, action);
     case PAGE_UNMOUNTED:
-      return Object.assign({}, state, {
-        newItems: [],
-        isUiDirty: false
-      });
+      return clearAnyUnsavedBreweriesFromState(state);
     case SET_ROW_AS_EDITABLE:
-      let stateCopy = Object.assign({}, state);
-      stateCopy.items.find(item=>item.brewery_id === action.rowId).isEditable = true;
-      return stateCopy;
+      return markBreweryAsEditable(state, action);
     case PUT_BREWERY_SUCCESS:
       // Clear out brewery state entirely, so list is refreshed from DB.
-      return Object.assign({}, state, {
-        items: []
-      });
+      return clearBreweriesFromState(state);
+    case DELETE_BREWERY_SUCCESS:
+      return removeBreweryFromState(state, action);
     default:
       return state;
-  }
-}
-
-export function breweryVisibilityFilter(state = VisibilityFilters.SHOW_ALL, action) {
-  switch (action.type) {
-    case SET_BREWERY_VISIBILITY_FILTER:
-      return action.filter;
-    default:
-      return state
   }
 }
