+++ beerweb/ui/app/styles/app.less
@@ -1,14 +1,11 @@
 @import '../../node_modules/elemental/less/elemental';
 
 .content-wrapper{
-  margin-top: 50px;
-  margin-right: 15%;
-  margin-left: 15%;
-  margin-bottom: 100px;
+  margin: 50px 15% 100px;
 }
 
 .table-wrapper{
-  margin: 0px;
+  margin: 0;
   color: @base;
 }
